refactor reporting error mapping module
